requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
azure-core==1.26.4
azure-cosmos==4.3.1
aiohttp